+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(ZeBook)
+
+test_check("ZeBook")